+++ index.html
@@ -111,7 +111,7 @@
 
      <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
      <script src="https://kit.fontawesome.com/7a65fe82ac.js" crossorigin="anonymous"></script>
-     <script type="module" src="./Js/script.js"></script>
+     <script type="module" src="./js/Main.js"></script>
 </body>
 
 </html>
